+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/mod.rs
@@ -1,7 +1,8 @@
 mod components;
-use components::{schemas::SchemasShape, ComponentsShapes};
 
 use crate::conversions::v3_0::to_rust_type::components::request_bodies::to_request_bodies_shape;
+use crate::conversions::v3_0::to_rust_type::components::schemas::SchemasShape;
+use crate::conversions::v3_0::to_rust_type::components::ComponentsShapes;
 use crate::conversions::{Result, ToRustType};
 use crate::targets::rust_type::{DataType, Modules};
 use openapi_types::v3_0::ComponentsObject;
